pyproject.toml
[project]
name = "bookly"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "asyncpg>=0.31.0",
    "fastapi>=0.128.0",
    "pydantic-settings>=2.12.0",
    "sqlmodel>=0.0.32",
]
